Fixed an issue within LitJsonFxJsonObjectConverter class where a value is coming from the webservice as float and was being parsed as an int.

+++ Assets/SocialPlay/Scripts/Web/LitJsonFxJsonObjectConverter.cs
@@ -201,7 +201,7 @@ public class LitJsonFxJsonObjectConverter : IServiceObjectConverter
             {
                 StoreItemDetail detail = new StoreItemDetail();
                 detail.propertyName = storeItemDetailArray[j]["Name"].ToString();
-                detail.propertyValue = int.Parse(storeItemDetailArray[j]["Value"].ToString());
+                detail.propertyValue = (int)float.Parse(storeItemDetailArray[j]["Value"].ToString());
 
                 //detail.invertEnergy = (bool)storeItemDetailArray[j]["InvertEnergy"];
